requirements.txt
alabaster==0.7.12
appdirs==1.4.3
argon2-cffi==19.2.0
asgiref==3.2.3
aspy.yaml==1.3.0
astroid==2.3.3
atomicwrites==1.3.0
attrs==19.3.0
Babel==2.8.0
backcall==0.1.0
black==19.10b0
certifi==2020.12.5
cffi==1.14.0
cfgv==3.0.0
chardet==3.0.4
Click==7.0
colorama==0.4.3
coreapi==2.3.3
coreschema==0.0.4
coverage==5.0.3
cryptography==3.0
decorator==4.4.1
defusedxml==0.6.0
dj-database-url==0.4.2
Django==2.2.20
django-allauth==0.41.0
django-anymail==7.0.0
django-appconf==1.0.4
django-autoslug==1.9.8
django-autoslug-iplweb==1.9.4.dev0
django-cities-light==3.8.1
django-coverage-plugin==1.6.0
django-crispy-forms==1.8.1
django-debug-toolbar==2.1
django-environ==0.4.5
django-extensions==3.0.8
django-filer==2.0.2
django-filter==2.3.0
django-js-asset==1.2.2
django-libs==2.0.3
django-localflavor==3.0.1
django-model-utils==4.0.0
django-mptt==0.11.0
django-oscar-paypal==2.0.0
django-polymorphic==3.0.0
django-redis==4.11.0
django-select2==7.7.1
django-stubs==1.4.0
djangorestframework==3.11.0
docutils==0.16
easy-thumbnails==2.7.1
entrypoints==0.3
factory-boy==2.12.0
Faker==4.0.0
flake8==3.7.9
gunicorn==20.0.4
identify==1.4.11
idna==2.8
imagesize==1.2.0
ipdb==0.12.3
ipython==7.12.0
ipython-genutils==0.2.0
isort==4.3.21
itypes==1.1.0
jedi==0.16.0
Jinja2==2.11.1
lazy-object-proxy==1.4.3
Markdown==3.2.2
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==8.2.0
mypy==0.761
mypy-extensions==0.4.3
nodeenv==1.3.5
oauthlib==3.1.0
packaging==20.1
parso==0.6.1
pathspec==0.7.0
pickleshare==0.7.5
Pillow==7.0.0
pluggy==0.13.1
pre-commit==1.21.0
progressbar2==3.53.1
prompt-toolkit==3.0.3
psycopg2==2.8.4
psycopg2-binary==2.8.4
py==1.8.1
pycodestyle==2.5.0
pycparser==2.19
pyflakes==2.1.1
Pygments==2.5.2
PyJWT==1.7.1
pylint==2.4.4
pylint-django==2.0.13
pylint-plugin-utils==0.6
pyparsing==2.4.6
PySocks==1.7.1
pytest==5.3.1
pytest-django==3.8.0
pytest-sugar==0.9.2
python-dateutil==2.8.1
python-slugify==4.0.0
python-stdnum==1.16
python-utils==2.4.0
python3-openid==3.1.0
pytz==2019.3
PyYAML==5.3
redis==3.3.11
regex==2020.1.8
requests==2.22.0
requests-oauthlib==1.3.0
rope==0.16.0
send-email==201212
six==1.15.0
snowballstemmer==2.0.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
Sphinx==2.3.1
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
sqlparse==0.3.0
swapper==1.1.2.post1
termcolor==1.1.0
text-unidecode==1.3
toml==0.10.0
tqdm==4.60.0
traitlets==4.3.3
tweepy==3.8.0
twilio==6.56.0
typed-ast==1.4.1
typing-extensions==3.7.4.1
Unidecode==1.1.1
uritemplate==3.0.1
urllib3==1.25.8
virtualenv==16.7.9
wcwidth==0.1.8
Werkzeug==0.16.0
whitenoise==5.0.1
wincertstore==0.2
wrapt==1.11.2
